requirements.txt
Pillow~=9.5.0
xlrd~=2.0.1
colorama~=0.4.6
python-dotenv~=1.0
